refactor(nvim): update plugins and CopilotChat.nvim branch
Several plugins have been updated in the lazy-lock.json file. The
CopilotChat.nvim plugin has been switched from a local directory to
the canary branch. This change will allow for the latest updates and
features to be used. Additionally, the selection parameter in the
Improve prompt has been commented out for further review and
adjustment.